Dropped support for t2-sim.
It did not work anymore and it makes no sense to
maintain that. So, just remove it to make some things
simpler.

As we have removed that, we can also remove the M3_MACHINE
thing, because this was only necessary to distinguish between
t2-chip and t2-sim. Maybe we add something similar later again,
but probably different.