kernel: fixed bug in sleeping in workloop.
Better wait at the beginning to ensure that we still have
things to wait for. Otherwise, it might happen that all
cores go to sleep and gem5 exits. Although it's not too
bad because we would have done that shortly after anyway,
it's nicer if the kernel shuts down properly.