Added a pager to VPEs to support completely paged execution.
That is, we can now map code and data from the file and add
anonymous mappings for the rest. In this way, all memory
of a VPE is paged. The pager is also passed to the child
VPE so that it can use the pager himself to e.g. map a file.

There are still some limitations and preliminary solutions,
though. For example, the heap is not implemented properly
(using mmap), clone doesn't work with paging, we map idle
in a very hacky way into the address space and so on.