gem5: use boot modules to load initial apps.
gem5 is now storing the initial apps somewhere in DRAM and
tells the kernel where they are. The kernel is then loading
them by looking at its arguments and searching for a boot
module with that name. Afterwards, it sets up an address
space and performs all the mappings.

With this, all boot modules are paged. If one specifies a
pager for VPEs, they are paged, too. However, if you don't
we are still using the old solution of letting gem5 map
a large range of memory on every PE.

The kernel is also responsible for putting idle on all
cores now. This removes the nasty hack in VPE::exec.