Added a backend to Env to distinguish between kernel and apps.
This is a much cleaner approach than the old way, where we
checked the core id or arguments to see who we are. Now,
we have an init function for the environment. libm3 provides
a default implementation that is intended for user applications.
The kernel overwrites this function to provide a different
implementation for the backend, suitable for the kernel.

This is intended to be the first step towards a separation
of libm3 into a basic library that is used both by the kernel
and user applications and a library on top, only used by
applications.