kernel: don't use libm3's {Send,Recv,Mem}Gate, RecvBuf and VPE.
This is the next step towards getting rid of most of libm3
in the kernel. Removing the gates, RecvBuf and VPE is a good
idea because:
1. Many features are not usable in the kernel. For example,
   creation of SendGates, revoking them and so on.
2. At the same time, many features are also not required in
   the kernel. For example, the EPMux is not necessary
   because the kernel can freely configure its EPs. Furthermore,
   some fields of the gates are not required.
3. Using these abstractions in the kernel did actually not make
   it easier, but more complicated and feels like a big hack.

However, the kernel does definitely want to use the GateStream
for marshalling and unmarshalling. This in turn depends on
SendGate, RecvGate and so on. Thus, we made GateStream more
general by making the gates template parameters (previous
commit) and provide our own versions of them in the kernel.
These are tuned for the kernel to be minimal and only contain
functionality that can actually be used in the kernel.