Added file descriptors and mountspace.
That is, VPEs now have file descriptors and a mountspace. The
classes FileTable and MountSpace are just containers that
can be accessed from VPE. On top of that, VFS and other classes
provide ways to mount filesystems, open files and so on.
This enables us to manipulate file descriptors and mounts of
child VPEs before starting them by just adding/removing pointers
to/from the containers.