Integraded pipes into the new file descriptor concept.
The current implementation has still a limitation, though.
Because a pipe is only usable by exactly the VPEs specified
at construction. That means, inheriting the pipe to a child
does not work, which is bad because applications should not
care which file implementation they are using.

But we can fix that in the future with a different concept for
pipes.