Reworked logging and use std streams instead of Serial.
It became more and more annoying that one had to rebuild
the entire code base whenever a logging flag was changed.
Thus, I've now moved PANIC out of it and split Log.h into
multiple files. One for the kernel, one for the library
and one for services.

Furthermore, the applications are now using cin, cout and
cerr instead of Serial, which moves us one step closer
to having a really usable OS.