Added IndirectPipe.
The advantage over DirectPipe is its flexibility, because it
supports multiple readers and writers and can be used from
everywhere. That is, not only between two preassigned PEs,
but from arbitrary PEs.
The disadvantage is that we need a central component and
can't just communicate directly between two PEs. To make
that scalable, we need one component per pipe. This has
not been done yet, but the plan is to let the pipe service
clone itself into a new VPE if a new pipe is created.