gem5: fixed bug in credit handling.
We have to treat unlimited credits special in the DTU by
not increasing/decreasing credits in this case. Because
the kernel needs that in order to reuse the send EP with
which he talks to services. Otherwise, they will reply
and thus give credits back, which increased the
credits on a potentially already changed EP.