gem5: allocate and enforce receive buffers in readonly memory.
This is achieved by a rw-barrier register in the DTU and by letting
the DTU disallow write accesses above that barrier. The DTU does
also prohibit to use a DTU read transfer into that region.

The kernel enforces that all receive buffers are above that
barrier to prevent that applications can change the header
before replying to a message.

This also helps us to treat receive buffers special. Because we
can't use techniques like COW, swapping etc. for them without
risking that communication is no longer possible in some cases.