kernel: allow overmap in createmap syscall.
This avoids that when changing permissions (e.g. for COW), we
need to first revoke the entire mapping and then recreate it
with the new permissions.

At the moment, the kernel does not change the physical memory
that is mapped there, but only the permissions. I don't know
if we need that in the future.