kernel: send close for server-created sessions, too.
If the server creates the session, the problem is that he owns
the root-capability. I.e., if the client revokes the session,
there is still the reference of the server left, so that the
kernel did not send the close message.

Therefore, I changed it, so that the close message is now send
when the next to last reference is gone for server-created
sessions. Note that the server still owns the session capability
and might also reuse that for a different client, if he likes.
He will then just receive another close message if the next to
last reference is gone again.