Improved flexibility of memory handling.
Previously, we had a fixed PE per platform, defined with a
constant, which hosted the main memory.

In the future, we want to support multiple memory modules and
thus, need to have more flexibility. Therefore, I rewrote the
whole main memory handling.

Now, we have platform-independent classes (MainMemory,
MemoryModule and MemoryMap), which are used to allocate and
free main memory. The Platform class inits MainMemory by
adding the corresponding MemoryModules to it.