Extend heap dynamically in kernel and use large heap for apps.
That is, the kernel uses a small initial heap, setup by gem5
(i.e., bootloader or so). This will be extended dynamically by
mapping more frames contiguously behind that.

For applications, we don't extend the heap yet. We could do that
later, if we want to, but currently this is not necessary,
because we just allocate a large virtual region for the heap
(64 MB atm) and map memory there dynamically (we don't need to
do anything for that).

Of course, this only happens if paging is available. If not,
we just use all available internal memory for the heap.