gem5: fixed a starvation bug in pagefault handling.
Since data transfers might cause pagefaults, which in turn
might require data transfers to be resolved (e.g., updating
PTEs by the kernel), we need to make sure that we can always
do that.

Thus, we can now specify that memory requests should not cause
pagefaults (if they do, they are aborted with an error). The
kernel does always specify that and the pager for the memory
cap to the virtual memory of the paged VPE.

This allows us to have a dedicated transfer buffer for transfers
that can't cause pagefaults. These will always finish eventually
without requiring other transfers and thus, we have no starvation
issue anymore.