The memory PEs don't need to be first.
We actually know already where the Env for the kernel
is stored, so just use that information.

With the current way we still need to have compute PEs
contiguous, though (because of the way we iterate over them).
If we want to mix them arbitrarily, we need to add an
indirection (translation table) or make the iteration more
complicated.