Removed virtual methods from Treap.
Since they incur some overhead, it's better to get rid of them
in performance-critical code. That means, that we can no longer
override the matches method of the Treap to specify how exactly
we find something. But this is not really required, because all
elements in one tree do it in the same way. Thus, shadowing the
method is enough.

Note that a consequence of this change is that every capability
needs to have a length property now, not only map caps.