kernel: introduced a slab allocator.
It is used for the capabilities and kernel objects at the moment to
optimize their allocation and deletion. The slab allocator is based
on the heap, i.e., requests memory from the heap.

To allow extending and shrinking slabs, every slab keeps a doubly linked
list of pools. However, currently, we do not shrink slabs to prevent
that we extend and shrink back and forth.