VPE: added start and stop operations.
The start operation existed before, but was not public. To use
non-programmable cores properly, we need both a start and a stop
operation in the VPE interface. Start is necessary, because if they
have VM support, we need to setup the address space and so on. Stop
is necessary, because they don't exit on their own.