gem5: improved rctmux to be more clear.
Now we use a separate program for idling instead of doing that in
rctmux. This allows us to always leave interrupts disabled in rctmux,
avoids a second stack and lets us do most of the stuff in C++.

Furthermore, the kernel does now initialize the address space after
the reset, so that we are sure about the PE's state.