gem5: improved msg reception.
The DTU uses a message array instead of a ringbuffer now and keeps
track of occupied and unread slots in that array. This allows us to keep
messages in the buffer until we reply to them, even if we handle other
messages in between.

That does also mean that we now have to specify the message to reply on,
which required some changes.