kernel: fixed service calls.
They are now delayed if the service VPE is currently suspended. To
ensure that we don't do a thread switch at arbitrary points, we program
a timeout to perform the switch from the workloop in a consistent state.
Additionally, this is done when calling received_reply, but we are in
a consistent state at that point, too.