Enabled tracing on gem5 as well and fixed bugs.
We use the tracing infrastructure on gem5 to track function calls. The
rest of the events are already visible in the gem5.log.

The gem52otf tool has been extended to support two different modes:
the "pes" mode displays the events happening between PEs, while the
"vpes" mode displays events between VPEs, i.e., the VPEs are the
processes. The former shows which VPE is running on which PE at what
points in time and the latter shows function calls.