Optimized syscall performance.
The compiler is unfortunately not smart enough to optimize the
GateStream operators as if we had hand written the store
operations. Using a struct and filling it by hand is thus better
performance-wise.

Additionally, we do not want to use the SendGate::send method,
because it performs several checks that are unnecessary in this case.
Because syscalls do never fail and the EP is never multiplexed.