kernel: migrate ready VPEs if a we have a free PE.
That is, whenever we context switch and the VPE is still ready,
we check whether another PE is available, which could continue
the execution.

We also need to update the idle-report flag if we migrate a VPE
to a PE, which hosts just one VPE atm. By setting the flag and
waking the PE up, the running VPE will either just continue
without any disruption until it wants to idle or if it already
idled, it will recheck the idle-report flag and then notify the
kernel about it.