kernel: enable reporting if we have any ready VPE.
That is, we do no longer do that on every PE individually, but keep
track of the number of global ready VPEs. If there is one, we enable
reporting on all (muxable) PEs. If a VPE yields, we check first if
there is a ready VPE on the same PE and if not, we steal a ready VPE
from a different PE.

This approach is much better, because we do not run into the problem
that a VPE on one PE is idling, but not reporting that to the kernel,
while another VPE is ready on a different PE.