hashaccel: create VPE for accelerator in client.
That is, we do no longer need a service to access the accelerator,
but the client is simply creating a VPE for the accelerator. This
provides the illusion that the client has exclusive access to the
accelerator. Shortly, we will support context switching on the
accelerator, so that multiple clients can use it without knowing of
each other.