gem5: let the DTU report whether idle was active.
That is, we do no longer let the application and rctmux try to
determine whether the application was idling. Instead, the DTU
remembers whether the last IRQ waked up the core. The kernel can
thus simply look at the saved register state to check whether
it should be blocked or not.