kernel: changed activate syscall.
Previously, we replaced the old capability with a new capability on
a given EP. I've decided to handle the race condition between credit
spending and activate on gem5 by using a new invalidate-EP command
instead of introducing cmpxchg as on the host. That is, we first try to
invalidate the EP, which fails if credits are missing. Afterwards, we
write the new content to the EP.

The syscall interface has also been changed accordingly. On the host, we
currently do not enforce that no credits are missing.