Made RecvBuf a capability.
Previously, we receive buffers were special, because we had a specific
syscall for detaching and simply specified the EP id in some cases
instead of what was actually meant: the receive buffer at that EP.

Now, RecvBuf is a capability as everything else. It is created with
createrbuf and destroyed via revoke. Furthermore, createrbuf does only
create a logical receive buffer, for which SendGates can be created.
To send messages to it, it needs to be activated first. This allows us
to let VPE A create a RecvBuf and a SendGate for it and then delegate
the RecvBuf to VPE B. VPE B can then activate it, so that VPE A can send
messages to VPE B using the SendGate. Of course, a RecvBuf can only be
activated by one VPE.