kernel: improved handling of service revokes.
Previously, we failed if there were outstanding client requests for
the service and let the service handle them first and retry the revoke
afterwards (of course, after preventing new requests). This was not a
really good solution.

The new solution is to just throw away the outstanding requests on the
service side and answer them with an error. The kernel will simply
activate the receive buffer again, if it's currently activated, which
resets, among other things, the occupied and unread bitmask. Thus, all
potentially outstanding messages are dropped.

The consequence is also that revokes can never fail, which simplifies
things (apart from specifying an invalid VPE cap or so).