Removed RecvGate.
The flexiblity it provided is actually not necessary and understanding
the purpose of having both RecvBuf and RecvGate was always hard. Thus,
I've decided to remove RecvGate and move the reply and wait operation
to RecvBuf. Furthermore, one can now start/stop a RecvBuf. Starting
registers a callback function and starts to listen to received messages.