Changed hash accelerator implementation.
We now use a start-update-finish API, which required some changes.
Currently, the service-version uses one accelerator-VPE per client and
lets the kernel context switch the accelerator in that way. This is a
bit strange when we want to compare it to the direct-version based on
context switching the accelerator. But the alternative would be to let
the accelerator offer an API to load/unload the state and let the
service use that. Since I do not want to make the accelerator more
complicated, this is the better option, I think.