VFS: allow to retrieve next memcap from File.
That is, we can call {read,write}_next on a File object, which
will give us a memory capability, offset within that memcap and
the its length. This allows us to, e.g., pass the memory cap to
an accelerator to let him directly read/write from/to that File.