hash: made the accelerator more autonomous.
Previously, we wrote the data to hash into the accelerator. With this
commit, we configure a memory EP at the accelerator and let the him
read the data autonomously from there. This improves the performance
quite a bit, because SW is not involved anymore, most of the time. And
we can let the potentially energy hungry general purpose core, that uses
the accelerator, go into a low power state.