SerialFile: don't hand out virtual memcap.
The hash accelerator does not support that. Unfortunately, it does
not work with the non-auto API of the hash accelerator now. But this
is not meant to be really used, but only to do a performance
comparison, so that it's okay that the serial line can't be used
for that.