Changed write_next API to begin and commit.
For writing, we need to first allocate space, but might not know
how much we actually want to write. Because for example, the
amount of written data might depend on the input and the algorithm
of an accelerator.

Thus, I've changed the API to allocate space first, but let us decide
later how much we actually use of that space. This already worked for
regular files, but not for pipes.