gem5: added protection for regs with MMU.
We have now 3 privilege levels in the DTU: master, privileged and
user. Master is used for the M3 kernel, privileged means that one
can access the "privileged registers". Accessing them allows to
handle requests from the kernel and translation requests from the
DTU.

As a second layer of protection, we map the privileged registers
to a separate page and only give the SW in the privileged CPU mode
access to it via MMU-based paging.