Don't free selectors.
For the kernel it's just a key for the Treap, so that the order
of the selector does not matter. Currently, it's a 32-bit
integer, but we can easily extend that to 64 or so. For practical
purposes, we can consider it unlimited (we can allocate 1000000
selectors per second for 500000 years ;)).