rustm3: close files via method, not Drop.
Using Drop does not work well here, because if we inherit files to
child VPEs, references are increased. To keep memory safety, I
decided to properly count the references and close the files
explicitly on FileTable::remove. This has the downside, that
"invalid" files stick around if somebody still has a reference.