fstrace: determine buffer size from trace.
Previously, we simply set MaxBufferSize sufficiently high to cover
all sizes. Since leveldb requires 32K and we use this buffer size
in sendfile, the performance of sendfile sucks if the buffer size
is larger than the I/O cache size. Thus, we use 8K for MaxBufferSize
again and determine the actual size of the buffer based on the
trace.