kernel: use short report time for services.
The rational is that if a service shares a PE with others, it is
only activated on demand and we don't want to wait just in case
another request arrives by accident. Hence, the kernel instructs
services to report idling immediately in case there is somebody
else ready to execute.