kernel: don't count waiting for groups members.
That is, if VPE A forwards a message or memory request to VPE B
and both VPEs are in the same VPE group, this is not considered
as "waiting" for VPE A. The rational is that groups are scheduled
simultaneously anyway. Therefore, we do want to schedule VPE A as
soon as VPE B is scheduled independent of whether there is a
forward syscall running.